python_backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
